You can no longer delete gid == 1
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/branches/1.2.x@4149 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ admin/listgroups.php
@@ -98,12 +98,16 @@ include_once("header.php");
                     echo $themeObject->DisplayImage('icons/system/edit.gif', lang('edit'),'','','systemicon');
                     echo "</a></td>\n";
                     }
-				if ($remove)
+				if ($remove && $onegroup->id != 1)
 				    {
-					echo "<td class=\"icons_wide\"><a href=\"deletegroup.php?group_id=".$onegroup->id."\" onclick=\"return confirm('".lang('deleteconfirm', $onegroup->name)."');\">";
-                    echo $themeObject->DisplayImage('icons/system/delete.gif', lang('delete'),'','','systemicon');
-                    echo "</a></td>\n";
-                    }
+				      echo "<td class=\"icons_wide\"><a href=\"deletegroup.php?group_id=".$onegroup->id."\" onclick=\"return confirm('".lang('deleteconfirm', $onegroup->name)."');\">";
+				      echo $themeObject->DisplayImage('icons/system/delete.gif', lang('delete'),'','','systemicon');
+				      echo "</a></td>\n";
+				    }
+				else
+				  {
+				    echo '<td class="icons_wide">&nbsp;</td>'."\n";
+				  }
 				echo "</tr>\n";
 
 				($currow == "row1"?$currow="row2":$currow="row1");
